requirements.txt
audioread==3.0.0
brotlipy==0.7.0
decorator==5.1.1
filelock==3.12.4
fsspec==2023.9.1
huggingface-hub==0.17.2
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.40.1
msgpack==1.0.5
numba==0.57.1
numpy==1.24.4
packaging==23.1
pandas==2.1.0
Pillow==9.4.0
platformdirs==3.10.0
pooch==1.7.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
soundfile==0.12.1
soxr==0.3.6
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.66.1
transformers==4.52.4
tzdata==2023.3
